refactor(sequelize-data-provider): convert to TS
