--- packages/sequelize-data-provider/src/index.js
@@ -1,5 +1,0 @@
-const sequelizeDataProvider = require('./sequelizeDataProvider');
-
-module.exports = {
-  sequelizeDataProvider
-};
