+++ packages/sequelize-data-provider/src/index.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export { default as sequelizeDataProvider } from './sequelizeDataProvider';
