+++ packages/sequelize-data-provider/src/sequelizeDataProvider.test.ts
@@ -1,9 +1,9 @@
 /* eslint-disable no-unused-vars */
-const { hacker, random } = require('faker');
-const { omit } = require('lodash');
-const sequelizeModel = require('@venncity/sequelize-model');
-const sequelizeDataProvider = require('./sequelizeDataProvider');
-const models = require('./../../../test/model');
+import { hacker, random } from 'faker';
+import { omit } from 'lodash';
+import sequelizeModel from '@venncity/sequelize-model';
+import sequelizeDataProvider from './sequelizeDataProvider';
+import models from '../../../test/model';
 
 sequelizeModel.sq.init(models);
